Allow users to remove themselves from a squad

+++ showup/views.py
@@ -91,27 +91,34 @@ def edit_squad(request, id):
             # Get my squad.
             my_squad = request.user.squad
 
-            # Get their squad and their members.
-            try:
-                their_squad = CustomUser.objects.get(email=request.POST["email"]).squad
+            if "add" in request.POST:
 
-                # We are already in the same squad.
-                if my_squad.id == their_squad.id:
+                # Get their squad and their members.
+                try:
+                    their_squad = CustomUser.objects.get(email=request.POST["email"]).squad
+
+                    # We are already in the same squad.
+                    if my_squad.id == their_squad.id:
+                        # TODO: Output some sort of message.
+                        return render(request, "edit_squad.html", {"form": form})
+
+                    their_members = CustomUser.objects.filter(squad=their_squad)
+                except CustomUser.DoesNotExist:
                     # TODO: Output some sort of message.
                     return render(request, "edit_squad.html", {"form": form})
 
-                their_members = CustomUser.objects.filter(squad=their_squad)
-            except CustomUser.DoesNotExist:
-                # TODO: Output some sort of message.
-                return render(request, "edit_squad.html", {"form": form})
+                # Merge squads.
+                for member in their_members:
+                    member.squad = my_squad
+                    member.save()
 
-            # Merge squads.
-            for member in their_members:
-                member.squad = my_squad
-                member.save()
+                # Delete their old squad.
+                Squad.objects.get(id=their_squad.id).delete()
 
-            # Delete their old squad.
-            Squad.objects.get(id=their_squad.id).delete()
+            elif "remove" in request.POST:
+                print(request.user.squad)
+                
+                
 
             return redirect(reverse("squad", kwargs={"id": id}))
         return render(request, "edit_squad.html", {"form": form})
